Allow multiple transitions with the same name
This works in Symfony's Workflow but laravel-workflow doesn't have a way to reference multiple transitions with the same name in the config file. This patch adds support for the alternative notation referenced in symfony/symfony#22558 that uses a `name` field rather than inferring the transition name from the key.

E.g.

```
	't1' => [
		'from' => 'a',
		'to' => 'b',
	]
```

becomes

```
	[
		'name' => 't1',
		'from' => 'a',
		'to' => 'b',
	]
```

which allows multiple transitions with the same name.